+++ src/filter-builder/filter-builder.d.ts
@@ -1,3 +1,9 @@
 import "../elements.js";
 
 export { RowanFilterBuilder } from "../../types/filter-builder/filter-builder.js";
+export type {
+  RowanFilter,
+  RowanFilterField,
+  RowanFilterFieldType,
+  RowanFilterOperator,
+} from "../../types/filter-builder/filter-builder.js";